Classifier/requirements.txt
# Core Streamlit & UI
streamlit
Pillow
flask               # For API server

# Image/PDF Processing
opencv-python-headless # For CV2 image operations
pdf2image            # For handling PDF files
easyocr              # For OCR text extraction
numpy                # For numerical operations (used with CV2, EasyOCR)

# Text Matching
fuzzywuzzy[speedup]  # For fuzzy text matching

# Langchain & AI Models
langchain
langchain-google-genai
python-dotenv        # For loading environment variables (like API keys)
chromadb             # Vector store used by the classifier

# Pydantic for data validation (used in utils.py)
pydantic

# Werkzeug for secure file handling in API
werkzeug
